pyproject.toml

[tool.black]
# Set line length explicitly.
# See https://black.readthedocs.io/en/stable/the_black_code_style.html#line-length for more details
exclude = "/*migrations*/"
line-length = 88

[tool.isort]
profile = "black"
skip_glob = "**/migrations/**"

[tool.poetry]
authors = ["Michael Sherman <shermanm@uchicago.edu>"]
description = ""
name = "portal"
version = "0.1.0"

[tool.poetry.dependencies]
Pillow = "^8.1.0"
PyYAML = "^5.3.1"
bibtexparser = "<1.2.0"
celery = "^5.0.5"
django = "~1.11.0"
django-angular = "^2.2"
django-bootstrap3 = "^12"
django-ckeditor = "^6.0.0"
django-classy-tags = "^1"
django-cms = "~3.6.0"
django-csp = "^3.7"
django-filer = "<2.0.0"
django-impersonate = "<1.5"
django-markdown-deux = "^1.0.5"
django-meta = "<2.0.0"
django-parler = "^2.0,<2.2"
django-pipeline = "^2.0.5"
django-polymorphic = "<3.0.0"
django-reversion = "^3.0,<3.0.8"
django-sekizai = "<2.0.0"
django-simple-captcha = "^0.5"
django-termsandconditions = "<2.0"
django-webpack-loader = "^0.7.0"
djangocms-attributes-field = "<2.0.0"
djangocms-blog = {git = "https://github.com/ChameleonCloud/djangocms-blog", rev = "chameleon/1.0.x"}
djangocms-column = "^1.11.0"
djangocms-file = "<3.0.0"
djangocms-flash = "^0.3.0"
djangocms-inherit = "^0.2.2"
djangocms-link = "<3.0.0"
djangocms-picture = "<3.0.0"
djangocms-style = "<3.0.0"
djangocms-teaser = "^0.2.0"
djangocms-text-ckeditor = "^3.10.0"
djangocms-video = "<3.0.0"
djangorestframework = "~3.11.2"
dynamic-rest = "^2.0.0"
httpretty = "^1.0.5"
kombu = "^5.0.2"
mock = "^4.0.3"
mozilla-django-oidc = {git = "https://github.com/chameleoncloud/mozilla-django-oidc"}
mysqlclient = "^1.0"
ply = "^3.11"
pytas = {git = "https://github.com/ChameleonCloud/pytas"}
python = "^3.6"
python-blazarclient = "2.2.1"
python-cinderclient = "^7.2.0"
python-glanceclient = "2.17.1"
python-keycloak-client = {git = "https://github.com/ChameleonCloud/python-keycloak-client"}
python-keystoneclient = "3.21.0"
python-memcached = "^1.59"
python-novaclient = "15.1.1"
python3-openid = "^3.2.0"
pytz = "^2020.5"
redis = "^3.5.3"
requests = "^2.25.1"
rt = "^2.0.1"
six = "^1.15.0"
slimit = "^0.8.1"
sqlparse = "^0.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pycodestyle = "^2.6.0"
pydocstyle = "^5.1.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
